+++ .coveragerc
@@ -1,14 +1,17 @@
 [run]
-relative_files = True
+source = src
+
+omit = */_version.py
 
 [report]
-omit =
-    */run.py
-    */python?.?/*
-    */venv/*
-    */site-packages/*
-    */tests/*
-    *__init__*
-
-exclude_lines =
-    if __name__ == '__main__':
+exclude_also =
+    from ._version
+    def __repr__
+    if self.debug:
+    if settings.DEBUG
+    raise AssertionError
+    raise NotImplementedError
+    if 0:
+    if __name__ == .__main__.:
+    if TYPE_CHECKING:
+    @(abc\.)?abstractmethod
