+++ .github/workflows/run-python-script.yml
@@ -12,15 +12,16 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-    - uses: actions/checkout@v3
+    - uses: actions/checkout@v4
     - name: Set up Python 3.x
-      uses: actions/setup-python@v4
+      uses: actions/setup-python@v5
       with:
         python-version: "3.x"
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+        pip install .
     - name: Run script
       env:
         HDX_SITE: ${{ vars.HDX_SITE }}
@@ -28,7 +29,7 @@ jobs:
         PREPREFIX: ${{ secrets.PREPREFIX }}
         USER_AGENT: ${{ secrets.USER_AGENT }}
       run: |
-        python run.py
+        python -m hdx.scraper.wfp.hungermap
     - name: Commit updated data bundle
       if: success()
       uses: stefanzweifel/git-auto-commit-action@v4
