+++ .pre-commit-config.yaml
@@ -1,0 +1,31 @@
+default_language_version:
+    python: python3.12
+repos:
+  - repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v5.0.0
+    hooks:
+      - id: trailing-whitespace
+      - id: end-of-file-fixer
+      - id: check-ast
+  - repo: https://github.com/astral-sh/ruff-pre-commit
+    rev: v0.9.10
+    hooks:
+      # Run the linter.
+      - id: ruff
+        args: [ --fix ]
+      # Run the formatter.
+      - id: ruff-format
+  - repo: https://github.com/astral-sh/uv-pre-commit
+    rev: 0.6.5
+    hooks:
+      # Run the pip compile
+      - id: pip-compile
+        name: pip-compile requirements.txt
+        files: pyproject.toml
+        args: [ pyproject.toml, --resolver=backtracking, --upgrade, -q,
+                -o, requirements.txt ]
+      - id: pip-compile
+        name: pip-compile requirements-test.txt
+        files: pyproject.toml
+        args: [ pyproject.toml, --resolver=backtracking, --upgrade, -q,
+                --extra, test, -c, requirements.txt, -o, requirements-test.txt ]
