+++ README.md
@@ -13,7 +13,7 @@ For the script to run, you will need to have a file called .hdx_configuration.ym
     hdx_key: "XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX"
     hdx_read_only: false
     hdx_site: prod
-    
+
  You will also need to supply the universal .useragents.yml file in your home directory as specified in the parameter *user_agent_config_yaml* passed to facade in run.py. The collector reads the key **hdx-scraper-wfp-hungermap** as specified in the parameter *user_agent_lookup*.
- 
+
  Alternatively, you can set up environment variables: USER_AGENT, HDX_KEY, HDX_SITE, TEMP_DIR, LOG_FILE_ONLY
