+++ pytest.ini
@@ -1,0 +1,4 @@
+[pytest]
+pythonpath = src
+addopts = "--color=yes"
+log_cli = 1
