+++ ruff.toml
@@ -1,0 +1,15 @@
+exclude = ["_version.py"]
+
+[lint]
+# List of rules: https://docs.astral.sh/ruff/rules/
+select = [
+  "E",   # pycodestyle - default
+  "F",   # pyflakes - default
+  "I"    # isort
+]
+ignore = [
+    "E501" # Line too long
+]
+
+[lint.isort]
+known-local-folder = ["hdx"]
