+++ src/hdx/scraper/wfp/hungermap/__init__.py
@@ -1,0 +1,1 @@
+from ._version import version as __version__  # noqa: F401
