+++ src/hdx/scraper/wfp/hungermap/config/hdx_resource_view_static.yaml
