--- test-requirements.txt
@@ -1,3 +1,0 @@
-pytest==8.3.5
-pytest-cov==6.0.0
--r requirements.txt
